requirements.txt
aiogram==3.14.0
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
python-decouple==3.8
openpyxl==3.1.5
redis==7.0.1
gunicorn==22.0.0
gevent==24.2.1
psycopg2-binary==2.9.9